refactor(api): extract plugins router (pure move onto the service)
Move the 25-route /plugins family out of src/api_server.py into
src/plugins/routes.py, with the mutating handlers now thin calls into
PluginService (which owns the shared mutate -> persist -> reset sequence
and the private-member reaches; no _mask_sensitive / _update_status
access remains in route code).

Per the #1756 pattern, every handler resolves the names the suite
monkeypatches — PLUGIN_SYSTEM_AVAILABLE, get_plugin_registry,
get_config_manager, reset_display_service, reset_template_engine,
get_template_engine, get_page_service, get_settings_service, and the
plugin-options limits/caches — through src.api_server at call time, so
patch('src.api_server.<name>') keeps steering the moved handlers. The
options machinery (constants, caches, semaphore, helpers) deliberately
stays in api_server: test_plugin_options_route.py monkeypatches it as
api_server module state, and the moved options handler borrows
api_server's logger so the never-log-a-draft-credential caplog test
stays live. unmask_sensitive_values / reset_display_service /
reset_template_engine are re-exported as documented patch seams.

api_server temporarily re-exports the seven handler names
src/mcp_server.py still imports; the follow-up MCP commit repoints those
tools at PluginService and deletes the aliases. The now-unused
_sanitize_optional_plugin_id moved into the service as
sanitize_optional_plugin_id.

Response-shape goldens (plugins domain, 43 records) and the route
inventory golden pass byte-identical; plugin/API/MCP batteries green.

Part of #1849; part of #1757.
