+++ src/api_server.py
@@ -39,9 +39,14 @@ from .board_client import board_client_from_board_dict  # noqa: E402
 from .collections.models import is_collection_id  # noqa: E402
 from .collections.service import get_collection_service  # noqa: E402
 from .config import Config  # noqa: E402
-from .config_manager import get_config_manager, unmask_sensitive_values  # noqa: E402
+
+# Patch seams (issues #1756/#1757): no handler left in this module calls
+# unmask_sensitive_values / reset_display_service / reset_template_engine, but
+# the extracted routers resolve them through `src.api_server` at call time so
+# tests that patch `src.api_server.<name>` keep working.
+from .config_manager import get_config_manager, unmask_sensitive_values  # noqa: E402, F401
 from .devices import classify_dimensions, resolve_dimensions  # noqa: E402
-from .displays.service import get_display_service, reset_display_service  # noqa: E402
+from .displays.service import get_display_service, reset_display_service  # noqa: E402, F401
 from .main import DisplayService  # noqa: E402
 from .network.wifi import WiFiError, get_wifi_service  # noqa: E402
 from .pages.service import (  # noqa: E402
@@ -58,7 +63,7 @@ from .paths import get_data_dir  # noqa: E402
 # tests that patch `src.api_server.get_schedule_service` keep working.
 from .schedules.service import get_schedule_service  # noqa: E402, F401
 from .settings.service import VALID_OUTPUT_TARGETS, VALID_STRATEGIES, get_settings_service  # noqa: E402
-from .templates.engine import get_template_engine, reset_template_engine  # noqa: E402
+from .templates.engine import get_template_engine, reset_template_engine  # noqa: E402, F401
 from .templates.expressions import function_signatures  # noqa: E402
 from .text_to_board import text_to_board_array, wrap_message_text  # noqa: E402
 from .time_service import reset_time_service  # noqa: E402
@@ -187,27 +192,6 @@ def _is_host_allowed(host: str, allowed_hosts: list[str]) -> bool:
 
 
 # Hostnames are restricted to RFC 1123 labels (letters, digits, hyphens) and
-_PLUGIN_ID_RE = re.compile(r"^[a-z0-9_]+$")
-
-
-def _sanitize_optional_plugin_id(plugin_id: str | None) -> str | None:
-    """Validate optional plugin id from user input.
-
-    Accepts ``None`` (meaning "derive from repo name"), otherwise enforces
-    lowercase letters, digits, and underscores only.
-    """
-    if plugin_id is None:
-        return None
-    if not isinstance(plugin_id, str) or not plugin_id:
-        raise HTTPException(status_code=400, detail="plugin_id must be a non-empty string")
-    if not _PLUGIN_ID_RE.fullmatch(plugin_id):
-        raise HTTPException(
-            status_code=400,
-            detail="plugin_id may contain only lowercase letters, digits, and underscores",
-        )
-    return plugin_id
-
-
 # IPv4 dotted-quad notation.  This rejects exotic forms (URL-encoded chars,
 # ``user:pass@host``, schemes embedded in the host, etc.) before we ever try
 # to connect to a board over HTTP.
@@ -8597,427 +8581,6 @@ except ImportError:
     get_plugin_registry = None
 
 
-class PluginConfigRequest(BaseModel):
-    """Request body for plugin configuration updates."""
-
-    config: dict[str, Any]
-
-
-class PluginEnableRequest(BaseModel):
-    """Request body for enabling/disabling a plugin."""
-
-    enabled: bool
-
-
-@app.get("/plugins")
-async def list_plugins():
-    """
-    List all available plugins.
-
-    Returns plugins with their status, metadata, and whether they're enabled.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    plugins = registry.list_plugins()
-
-    # Add configuration status (masked)
-    config_manager = get_config_manager()
-    for plugin in plugins:
-        plugin_config = config_manager.get_plugin_config(plugin["id"])
-        if plugin_config:
-            plugin["configured"] = True
-            # Add masked config
-            plugin["config"] = config_manager._mask_sensitive(plugin_config)
-        else:
-            plugin["configured"] = False
-            plugin["config"] = {}
-
-    return {
-        "plugins": plugins,
-        "plugin_system_enabled": True,
-        "total": len(plugins),
-        "enabled_count": sum(1 for p in plugins if p.get("enabled", False)),
-    }
-
-
-@app.get("/plugins/variables/all")
-async def get_all_plugin_variables():
-    """
-    Get all template variables from enabled plugins.
-
-    Returns a combined view of all variables for the template editor.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        # Fall back to legacy variables
-        template_engine = get_template_engine()
-        return {
-            "variables": template_engine.get_available_variables(),
-            "max_lengths": template_engine.get_variable_max_lengths(),
-            "plugin_system_enabled": False,
-        }
-
-    registry = get_plugin_registry()
-
-    return {
-        "variables": registry.get_all_variables(),
-        "max_lengths": registry.get_all_max_lengths(),
-        "plugin_system_enabled": True,
-    }
-
-
-@app.get("/plugins/errors")
-async def get_plugin_errors():
-    """
-    Get any plugin load errors.
-
-    Returns errors from plugins that failed to load.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        return {"errors": {}, "plugin_system_enabled": False}
-
-    registry = get_plugin_registry()
-
-    return {"errors": registry.get_load_errors(), "plugin_system_enabled": True}
-
-
-@app.get("/plugins/registry")
-async def list_registry_plugins():
-    """
-    List all plugins available in the curated plugin registry.
-
-    Returns registry entries with their installation status.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    return {
-        "entries": registry.get_registry_entries(),
-        "plugin_system_enabled": True,
-    }
-
-
-@app.get("/plugins/updates")
-async def get_plugin_updates():
-    """
-    Return cached update availability for all installed external plugins.
-
-    Results are refreshed by a background task every 6 hours.  Call
-    ``POST /plugins/updates/check`` to trigger an immediate check.
-
-    ``blocked`` maps plugin ids to the reason an upstream commit was *not*
-    offered — currently only "the incoming manifest needs a newer FiestaBoard
-    core".  Those plugins appear in ``updates`` as ``False``; the reason is
-    what lets the UI say so rather than looking stuck.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    return {
-        "updates": registry.get_update_status(),
-        "blocked": registry.get_update_blocked_reasons(),
-    }
-
-
-@app.get("/plugins/{plugin_id}")
-async def get_plugin(plugin_id: str):
-    """
-    Get details for a specific plugin.
-
-    Returns the plugin's manifest, configuration, and status.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    manifest = registry.get_manifest(plugin_id)
-
-    if not manifest:
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    # STORED configuration, without the env-var overlay: this response feeds
-    # the settings form, and any value baked in here comes straight back in
-    # the next save — serving the overlay would freeze env values into
-    # config.json (#1864 review). Which keys are currently env-controlled is
-    # reported separately in env_overridden_keys.
-    config_manager = get_config_manager()
-    plugin_config = config_manager.get_plugin_config(plugin_id, include_env_overrides=False)
-    env_overridden_keys = sorted(config_manager.get_plugin_env_overrides(plugin_id)) if plugin_config else []
-
-    # Check for demo page (use flagship as the representative for backwards compat)
-    has_demo = manifest.demo is not None
-    demo_page_id = None
-    if has_demo:
-        page_service = get_page_service()
-        demo_page = page_service.get_demo_page(plugin_id, device_type="flagship") or page_service.get_demo_page(
-            plugin_id
-        )
-        if demo_page:
-            demo_page_id = demo_page.id
-
-    # Instance information
-    base_id, instance_label = registry.parse_instance_key(plugin_id)
-    instances = registry.list_instances(base_id) if not instance_label else []
-
-    return {
-        "id": plugin_id,
-        "name": manifest.name,
-        "version": manifest.version,
-        "description": manifest.description,
-        "author": manifest.author,
-        "icon": manifest.icon,
-        "category": manifest.category,
-        # "data" or "transition" -- the UI hides the enable toggle for
-        # transition plugins, which run whenever selected regardless of it.
-        "plugin_type": manifest.plugin_type,
-        "enabled": registry.is_enabled(plugin_id),
-        "config": config_manager._mask_sensitive(plugin_config) if plugin_config else {},
-        # Config keys whose live value currently comes from an env var (the
-        # values themselves are deliberately NOT in "config").
-        "env_overridden_keys": env_overridden_keys,
-        "settings_schema": manifest.settings_schema,
-        "variables": manifest.raw.get("variables", {}),
-        "max_lengths": manifest.max_lengths,
-        "env_vars": manifest.env_vars,
-        "documentation": manifest.documentation,
-        "has_demo": has_demo,
-        "demo_page_id": demo_page_id,
-        "instance_label": instance_label,
-        "base_plugin_id": base_id,
-        "instances": instances,
-    }
-
-
-@app.get("/plugins/{plugin_id}/manifest")
-async def get_plugin_manifest(plugin_id: str):
-    """
-    Get the full manifest for a plugin.
-
-    Returns the raw manifest data for UI rendering.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    manifest = registry.get_manifest(plugin_id)
-
-    if not manifest:
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    return manifest.raw
-
-
-@app.put("/plugins/{plugin_id}/config")
-async def update_plugin_config(plugin_id: str, request: PluginConfigRequest):
-    """
-    Update configuration for a plugin.
-
-    Args:
-        plugin_id: Plugin identifier
-        request: Configuration to apply
-
-    Example body:
-    {
-        "config": {
-            "api_key": "your-api-key",
-            "location": "San Francisco, CA",
-            "refresh_seconds": 300
-        }
-    }
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    # Check if plugin exists
-    if not registry.get_plugin(plugin_id):
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    # Resolve the "***" placeholders before anything sees the payload. Config
-    # goes out masked, so any client that echoes back what it read — the
-    # settings form, or an MCP client working from list_installed_plugins() —
-    # posts the sentinel where the secret used to be. Un-masking here rather
-    # than only inside ConfigManager keeps the mask out of the live plugin as
-    # well, whose validate_config()/on_config_change() would otherwise run
-    # against three asterisks (issue #1743).
-    config_manager = get_config_manager()
-    # Raw stored config, WITHOUT the env-var overlay: this value feeds the
-    # un-mask + persist path, and restoring "***" from the overlay would
-    # silently write an env secret into config.json (issue #1761).
-    stored_config = config_manager.get_plugin_config(plugin_id, include_env_overrides=False) or {}
-    config = unmask_sensitive_values(request.config, stored_config)
-
-    # Validate configuration against manifest schema
-    errors = registry.set_plugin_config(plugin_id, config)
-    if errors:
-        logger.error(f"Plugin '{plugin_id}' config validation failed: {errors}")
-        raise HTTPException(status_code=400, detail={"errors": errors})
-
-    # Save to config file
-    config_manager.set_plugin_config(plugin_id, config)
-
-    # Re-seed the LIVE registry config from the overlaid read. Persisting
-    # env-free is correct, but the validation call above also installed that
-    # env-free dict as the live config — killing a working env-supplied
-    # credential until restart (#1864 review). Named instances never have an
-    # overlay, so this is a no-op for them.
-    live_config = config_manager.get_plugin_config(plugin_id)
-    if live_config is not None and live_config != config:
-        reseed_errors = registry.set_plugin_config(plugin_id, live_config)
-        if reseed_errors:
-            # The stored config validated; only the env overlay can be at
-            # fault. Keep the env-free live config rather than failing the
-            # save the user just made.
-            logger.warning(
-                f"Env overlay for '{plugin_id}' failed validation after save: {reseed_errors} "
-                "— live config runs without the overlay until restart"
-            )
-
-    # Reset services to pick up new config
-    reset_display_service()
-    reset_template_engine()
-
-    logger.info(f"Plugin '{plugin_id}' configuration updated")
-
-    # Return masked config
-    updated = config_manager.get_plugin_config(plugin_id)
-
-    return {
-        "status": "success",
-        "plugin_id": plugin_id,
-        "config": config_manager._mask_sensitive(updated) if updated else {},
-    }
-
-
-@app.post("/plugins/{plugin_id}/enable")
-async def enable_plugin(plugin_id: str):
-    """
-    Enable a plugin.
-
-    Enables the plugin in both the registry and persists to config.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    if not registry.get_plugin(plugin_id):
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    # Enable in registry
-    success = registry.enable_plugin(plugin_id)
-    if not success:
-        raise HTTPException(status_code=400, detail=f"Failed to enable plugin: {plugin_id}")
-
-    # Persist to config
-    config_manager = get_config_manager()
-    config_manager.enable_plugin(plugin_id)
-
-    # Reset services
-    reset_display_service()
-    reset_template_engine()
-
-    logger.info(f"Plugin '{plugin_id}' enabled")
-
-    return {"status": "success", "plugin_id": plugin_id, "enabled": True}
-
-
-@app.post("/plugins/{plugin_id}/disable")
-async def disable_plugin(plugin_id: str):
-    """
-    Disable a plugin.
-
-    Disables the plugin in both the registry and persists to config.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    if not registry.get_plugin(plugin_id):
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    # Disable in registry
-    success = registry.disable_plugin(plugin_id)
-    if not success:
-        raise HTTPException(status_code=400, detail=f"Failed to disable plugin: {plugin_id}")
-
-    # Persist to config
-    config_manager = get_config_manager()
-    config_manager.disable_plugin(plugin_id)
-
-    # Reset services
-    reset_display_service()
-    reset_template_engine()
-
-    logger.info(f"Plugin '{plugin_id}' disabled")
-
-    return {"status": "success", "plugin_id": plugin_id, "enabled": False}
-
-
-@app.get("/plugins/{plugin_id}/data")
-async def get_plugin_data(plugin_id: str):
-    """
-    Fetch current data from a plugin.
-
-    Returns the plugin's latest data, formatted output, and status.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    if not registry.get_plugin(plugin_id):
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    if not registry.is_enabled(plugin_id):
-        raise HTTPException(status_code=400, detail=f"Plugin not enabled: {plugin_id}")
-
-    result = registry.fetch_plugin_data(plugin_id)
-
-    # Return 503 when plugin data is unavailable (e.g. not configured, auth failure)
-    # so monitoring (Grafana) and request log show it as an error for triage
-    if not result.available:
-        raise HTTPException(status_code=503, detail=result.error or "Plugin data not available")
-
-    return {
-        "plugin_id": plugin_id,
-        "available": result.available,
-        "data": result.data,
-        "formatted_lines": result.formatted_lines,
-        "error": result.error,
-    }
-
-
-@app.get("/plugins/{plugin_id}/variables")
-async def get_plugin_variables(plugin_id: str):
-    """
-    Get template variables exposed by a plugin.
-
-    Returns the variables schema for use in the template editor.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    manifest = registry.get_manifest(plugin_id)
-
-    if not manifest:
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    return {
-        "plugin_id": plugin_id,
-        "variables": manifest.raw.get("variables", {}),
-        "max_lengths": manifest.max_lengths,
-        "color_rules_schema": manifest.raw.get("color_rules_schema", {}),
-    }
-
-
 # ── Plugin remote options ────────────────────────────────────────────────────
 
 # Hard ceiling on how many options core will hand to the browser, whatever the
@@ -9081,22 +8644,6 @@ async def _bounded_options_call(work: Any, timeout: float) -> Any:
     return await asyncio.wait_for(asyncio.shield(task), timeout)
 
 
-class PluginOptionsRequestBody(BaseModel):
-    """Body for ``POST /plugins/{plugin_id}/options/{options_id}``.
-
-    POST rather than GET on purpose: ``parent`` holds arbitrary JSON, and
-    ``draft_config`` carries credentials that must never reach a URL, an
-    access log, or browser history.
-    """
-
-    parent: dict[str, Any] = Field(default_factory=dict)
-    query: str = ""
-    limit: int = 200
-    cursor: str | None = None
-    refresh: bool = False
-    draft_config: dict[str, Any] = Field(default_factory=dict)
-
-
 # Answers are cached per (plugin instance, provider, effective config, query)
 # so that typing in a search box does not hammer an upstream API. Bounded so a
 # long-lived process cannot accumulate one entry per keystroke forever.
@@ -9121,7 +8668,7 @@ def _plugin_options_cache_key(
     plugin_id: str,
     options_id: str,
     config_digest: str,
-    body: "PluginOptionsRequestBody",
+    body: Any,  # PluginOptionsRequestBody (moved to src/plugins/routes.py); duck-typed here
     limit: int,
 ) -> str:
     """Build the cache key for one options question.
@@ -9294,741 +8841,26 @@ def _fit_options_payload(payload: dict[str, Any]) -> dict[str, Any]:
     return payload
 
 
-@app.post("/plugins/{plugin_id}/options/{options_id}")
-async def get_plugin_options_endpoint(plugin_id: str, options_id: str, body: PluginOptionsRequestBody):
-    """Browse a plugin's upstream catalog to populate one settings field."""
-    from .plugins.base import OptionsRequest, OptionsUnavailable
-
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    if registry.get_plugin(plugin_id) is None:
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    manifest = registry.get_manifest(plugin_id)
-    if options_id not in _declared_options_ids(manifest):
-        raise HTTPException(
-            status_code=400,
-            detail=f"Plugin '{plugin_id}' does not declare options provider '{options_id}'",
-        )
-
-    limit = max(1, min(body.limit, PLUGIN_OPTIONS_MAX_RETURNED))
-    request = OptionsRequest(
-        options_id=options_id,
-        parent=body.parent,
-        query=body.query,
-        limit=limit,
-        cursor=body.cursor,
-    )
-
-    cache_seconds = _options_cache_seconds(manifest, options_id)
-
-    def _envelope(**overrides: Any) -> dict[str, Any]:
-        payload: dict[str, Any] = {
-            "plugin_id": plugin_id,
-            "options_id": options_id,
-            "options": [],
-            "has_more": False,
-            "cursor": None,
-            "total": None,
-            "error": None,
-            "cached": False,
-            "stale": False,
-            "cache_seconds": cache_seconds,
-        }
-        payload.update(overrides)
-        return payload
-
-    stored_config = dict(registry.get_plugin_config(plugin_id) or {})
-    # The form posts back "***" wherever a sensitive field used to be, so the
-    # draft has to be un-masked against what is stored or the plugin gets three
-    # asterisks as its API key and every lookup fails mid-setup.
-    draft_config = unmask_sensitive_values(body.draft_config, stored_config) if body.draft_config else None
-    if draft_config:
-        # Key names only, never values: a settings dialog is exactly where
-        # credentials leak into logs.
-        logger.debug(
-            "Options request for '%s/%s' carries draft config keys: %s",
-            plugin_id,
-            options_id,
-            sorted(draft_config),
-        )
-
-    effective_config = {**stored_config, **(draft_config or {})}
-    cache_key = _plugin_options_cache_key(plugin_id, options_id, _config_fingerprint(effective_config), body, limit)
-
-    if body.refresh:
-        _plugin_options_refresh_throttle(cache_key)
-    elif cache_seconds > 0:
-        entry = _plugin_options_cache_get(cache_key)
-        if entry is not None and (time.monotonic() - entry[0]) < cache_seconds:
-            return {**entry[1], "cached": True, "stale": False}
-
-    try:
-        # Never call the plugin inline: get_options() makes network calls, and
-        # blocking the event loop here would stall every other request in the
-        # process. (GET /plugins/{id}/data still does this; do not copy it.)
-        result = await _bounded_options_call(
-            lambda: registry.get_plugin_options(plugin_id, options_id, request, draft_config=draft_config),
-            PLUGIN_OPTIONS_TIMEOUT_SECONDS,
-        )
-    except TimeoutError as e:
-        logger.warning("Options provider '%s' timed out for plugin '%s'", options_id, plugin_id)
-        stale = _stale_options_payload(cache_key, reason=f"Options provider '{options_id}' timed out")
-        if stale is not None:
-            return stale
-        raise HTTPException(
-            status_code=504,
-            detail=f"Options provider '{options_id}' timed out",
-        ) from e
-    except NotImplementedError as e:
-        # The manifest promised a provider the class never implemented, or the
-        # plugin is a transition. 501 lets the widget degrade to a plain input.
-        raise HTTPException(status_code=501, detail=str(e)) from e
-    except KeyError as e:
-        # The registry could not build a sandbox: the plugin was uninstalled
-        # while its settings dialog was open. Still "no such plugin".
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}") from e
-    except OptionsUnavailable as e:
-        # Deliberately a 200. "No API key yet" is the *expected* state while
-        # the user is still filling the form in; the widget shows the reason
-        # inline next to the field instead of a failed-request toast.
-        return _envelope(error=str(e))
-    except Exception as e:
-        # The traceback (with the plugin's raw error text) is already in the
-        # server log; the client gets a static message so plugin exceptions
-        # cannot leak keys/URLs/paths (CodeQL py/stack-trace-exposure).
-        logger.exception("Options provider '%s' failed for plugin '%s'", options_id, plugin_id)
-        stale = _stale_options_payload(cache_key, reason="Options provider failed")
-        if stale is not None:
-            return stale
-        raise HTTPException(status_code=502, detail="Options provider failed") from e
-
-    options, truncated = _serialise_options(result.options, limit, plugin_id, options_id)
-    payload = _fit_options_payload(
-        _envelope(
-            options=options,
-            has_more=bool(result.has_more) or truncated,
-            cursor=_truncate(result.cursor, PLUGIN_OPTIONS_MAX_CURSOR_CHARS),
-            total=result.total,
-            error=result.error,
-        )
-    )
-
-    if cache_seconds > 0:
-        _plugin_options_cache_put(cache_key, payload)
-
-    return payload
-
-
-# ── Plugin Demo Pages ────────────────────────────────────────────────────────
-
-
-def _resolve_demo_device_type(demo: dict) -> str:
-    """Pick the demo device_type that matches the configured board.
-
-    Walks the user's configured boards in order and returns the first
-    device_type that the plugin actually ships a demo for. Falls back to
-    any device_type the plugin supports, then to "flagship" as a last
-    resort. See issue #942.
-    """
-    configured: list[str] = []
-    try:
-        board_settings = get_settings_service().get_board_settings()
-        for board in getattr(board_settings, "boards", []) or []:
-            dt = board.get("device_type") if isinstance(board, dict) else None
-            if dt and dt not in configured:
-                configured.append(dt)
-    except Exception:  # noqa: BLE001 — settings access must never break demo creation
-        logger.debug("Could not resolve configured device_type; using plugin default", exc_info=True)
-
-    for dt in configured:
-        if dt in demo:
-            return dt
-    if demo:
-        return next(iter(demo))
-    return "flagship"
-
-
-@app.get("/plugins/{plugin_id}/demo-page")
-async def get_plugin_demo_page(plugin_id: str, device_type: str = "flagship"):
-    """
-    Check whether a demo page exists for this plugin and device type.
-
-    Returns ``exists: true`` and the page id when one is found.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    manifest = registry.get_manifest(plugin_id)
-    if not manifest:
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    if manifest.demo is None:
-        return {"exists": False, "page_id": None, "has_demo_template": False}
-
-    has_demo_template = device_type in manifest.demo
-    page_service = get_page_service()
-    demo_page = page_service.get_demo_page(plugin_id, device_type=device_type)
-    return {
-        "exists": demo_page is not None,
-        "page_id": demo_page.id if demo_page else None,
-        "has_demo_template": has_demo_template,
-    }
-
-
-@app.post("/plugins/{plugin_id}/demo-page")
-async def create_plugin_demo_page(plugin_id: str, device_type: str | None = None):
-    """
-    Create (or recreate) the demo page for a plugin and device type.
-
-    When *device_type* is omitted, it is resolved from the configured board
-    settings (the first device type listed under Settings → Hardware), so a
-    Note board does not silently get a Flagship-sized demo page (issue #942).
-    If the plugin does not ship a demo template for the configured device,
-    we fall back to any device type it does support.
-
-    The demo page is a singleton per plugin + device type -- calling this endpoint
-    when a demo page already exists for that device type will delete the old one
-    and create a fresh copy.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    manifest = registry.get_manifest(plugin_id)
-    if not manifest:
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-
-    if manifest.demo is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Plugin '{plugin_id}' does not include a demo page template.",
-        )
-
-    resolved_device_type = device_type or _resolve_demo_device_type(manifest.demo)
-
-    demo_schema = manifest.demo.get(resolved_device_type)
-    if demo_schema is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Plugin '{plugin_id}' has no demo template for device type '{resolved_device_type}'.",
-        )
-
-    # Check that required settings are configured
-    settings_schema = manifest.settings_schema
-    required_fields = settings_schema.get("required", [])
-    if required_fields:
-        config_manager = get_config_manager()
-        plugin_config = config_manager.get_plugin_config(plugin_id) or {}
-        missing = [f for f in required_fields if f != "enabled" and not plugin_config.get(f)]
-        if missing:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Required settings not configured: {', '.join(missing)}. "
-                f"Configure them first before creating a demo page.",
-            )
-
-    page_service = get_page_service()
-    page, recreated = page_service.create_demo_page(plugin_id, demo_schema)
-
-    return {
-        "status": "recreated" if recreated else "created",
-        "page": page.model_dump(),
-    }
-
-
-# ── Plugin Instances ────────────────────────────────────────────────────────
-
-
-class PluginInstanceCreateRequest(BaseModel):
-    """Request body for creating a new plugin instance."""
-
-    label: str
-
-
-@app.get("/plugins/{plugin_id}/instances")
-async def list_plugin_instances(plugin_id: str):
-    """
-    List all instances of a plugin.
-
-    Returns the instances (excluding the base) for the given plugin.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    # Resolve base plugin id (strip instance label if present)
-    base_id, _ = registry.parse_instance_key(plugin_id)
-
-    if not registry.get_plugin(base_id):
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {base_id}")
-
-    instances = registry.list_instances(base_id)
-
-    return {
-        "plugin_id": base_id,
-        "instances": instances,
-        "total": len(instances),
-    }
-
-
-@app.post("/plugins/{plugin_id}/instances")
-async def create_plugin_instance(plugin_id: str, request: PluginInstanceCreateRequest):
-    """
-    Create a new instance of a plugin.
-
-    The new instance starts disabled with an empty configuration.
-    It can be configured and enabled independently via the standard
-    plugin config/enable endpoints using the compound key
-    ``{plugin_id}:{label}``.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    # Resolve base plugin id
-    base_id, _ = registry.parse_instance_key(plugin_id)
-
-    if not registry.get_plugin(base_id):
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {base_id}")
-
-    errors = registry.create_instance(base_id, request.label)
-    if errors:
-        raise HTTPException(status_code=400, detail="; ".join(errors))
-
-    compound_key = registry.make_instance_key(base_id, request.label)
-
-    config_manager = get_config_manager()
-    # The registry can come up without its named instances — an unreadable
-    # config.json, or a base plugin that failed to load, both leave the stored
-    # entries untouched but drop the live instances. The UI then shows nothing
-    # and the user re-adds the label by hand. Blindly persisting an empty
-    # config here used to overwrite their saved settings with
-    # `{"enabled": false}`, so the plugin fell back to manifest defaults.
-    # Adopt whatever is still on disk instead.
-    stored = config_manager.get_plugin_config(compound_key)
-    if stored:
-        errors = registry.apply_stored_config(compound_key, stored)
-        if errors:
-            logger.warning(
-                "Adopted stored config for re-created instance '%s' despite validation errors: %s",
-                compound_key,
-                errors,
-            )
-        if stored.get("enabled"):
-            registry.enable_plugin(compound_key)
-        logger.info("Re-created instance '%s' adopted its existing stored config", compound_key)
-    else:
-        # Persist empty config so the instance survives restarts
-        config_manager.set_plugin_config(compound_key, {"enabled": False})
-    # Re-creating an instance is an explicit user action — drop any
-    # deliberate-removal tombstone left by a prior delete (#1394).
-    config_manager.clear_plugin_removed(compound_key)
-
-    # Reset services so the new instance is available to templates immediately
-    reset_display_service()
-    reset_template_engine()
-
-    logger.info(f"Created plugin instance: {compound_key}")
-
-    # Report the normalized label — that is the instance the registry holds and
-    # the one `{{plugin:label.field}}` template references must use.
-    _, instance_label = registry.parse_instance_key(compound_key)
-
-    return {
-        "status": "success",
-        "plugin_id": base_id,
-        "instance_label": instance_label,
-        "instance_key": compound_key,
-        "message": f"Instance '{instance_label}' created for plugin '{base_id}'.",
-    }
-
-
-@app.delete("/plugins/{plugin_id}/instances/{instance_label}")
-async def delete_plugin_instance(plugin_id: str, instance_label: str):
-    """
-    Delete a plugin instance.
-
-    Removes the instance from the registry and its persisted configuration.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    # Resolve base plugin id
-    base_id, _ = registry.parse_instance_key(plugin_id)
-
-    errors = registry.delete_instance(base_id, instance_label)
-    if errors:
-        raise HTTPException(status_code=400, detail="; ".join(errors))
-
-    compound_key = registry.make_instance_key(base_id, instance_label)
-
-    # Remove persisted config and tombstone the compound key so a
-    # post-upgrade auto-restore cannot resurrect the deleted instance (#1394).
-    config_manager = get_config_manager()
-    config_manager.delete_plugin_config(compound_key)
-    config_manager.mark_plugin_removed(compound_key)
-
-    # Reset services
-    reset_display_service()
-    reset_template_engine()
-
-    logger.info(f"Deleted plugin instance: {compound_key}")
-
-    return {
-        "status": "success",
-        "plugin_id": base_id,
-        "instance_label": instance_label,
-        "instance_key": compound_key,
-        "message": f"Instance '{instance_label}' of plugin '{base_id}' deleted.",
-    }
-
-
-@app.post("/plugins/{plugin_id}/receive")
-async def receive_plugin_payload(plugin_id: str, request: Request):
-    """
-    Push a JSON payload to a plugin.
-
-    Allows external systems (CI pipelines, automations, etc.) to push data to
-    plugins that support incoming webhooks.  The plugin's ``receive_payload``
-    method is called with the parsed body, the raw request headers, and the
-    raw body bytes (for HMAC verification).
-
-    Returns 404 when the plugin is not found, 400 when it is not enabled or
-    the body is not valid JSON, 403 when the plugin rejects the request due to
-    a signature mismatch, and 405 when the plugin does not support receive.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    plugin = registry.get_plugin(plugin_id)
-    if not plugin:
-        raise HTTPException(status_code=404, detail=f"Plugin not found: {plugin_id}")
-    if not registry.is_enabled(plugin_id):
-        raise HTTPException(status_code=400, detail=f"Plugin not enabled: {plugin_id}")
-
-    raw_body = await request.body()
-    try:
-        body = json.loads(raw_body)
-    except Exception:
-        raise HTTPException(status_code=400, detail="Request body must be valid JSON") from None
-
-    headers = dict(request.headers)
-
-    try:
-        plugin.receive_payload(body, headers, raw_body=raw_body)
-    except NotImplementedError:
-        raise HTTPException(
-            status_code=405,
-            detail=f"Plugin '{plugin_id}' does not support receive",
-        ) from None
-    except PermissionError as exc:
-        raise HTTPException(status_code=403, detail=str(exc)) from exc
-    except ValueError as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    return {"status": "ok"}
-
-
-# ── External Plugin Management ──────────────────────────────────────────────
-
-
-class ExternalPluginInstallRequest(BaseModel):
-    """Request body for installing an external plugin."""
-
-    repository: str
-    plugin_id: str | None = None
-    branch: str = ""
-
-
-@app.post("/plugins/registry/{plugin_id}/install")
-async def install_registry_plugin(plugin_id: str):
-    """
-    Install a plugin from the curated registry by its id.
-
-    The install shells out to ``git`` (up to 120 s) and then imports the
-    plugin package, so it runs in a worker thread — inline it would seize the
-    event loop and freeze every other request for the whole clone (#1750).
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    errors = await asyncio.to_thread(registry.install_from_registry, plugin_id)
-
-    if errors:
-        raise HTTPException(status_code=400, detail="; ".join(errors))
-
-    return {
-        "status": "success",
-        "plugin_id": plugin_id,
-        "message": f"Plugin '{plugin_id}' installed from registry.",
-    }
-
-
-@app.post("/plugins/install")
-async def install_external_plugin(request: ExternalPluginInstallRequest):
-    """
-    Install a plugin from a public git repository URL.
-
-    The repository does not need to follow the ``fiestaboard-plugin--``
-    naming convention (that requirement only applies to registry plugins).
-
-    The clone runs in a worker thread so a slow or unreachable remote cannot
-    block the event loop (#1750).
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    safe_branch = request.branch or ""
-    if safe_branch:
-        from .plugins.sources import _validate_git_ref
-
-        _ok, _err = _validate_git_ref(safe_branch)
-        if not _ok:
-            raise HTTPException(status_code=400, detail=_err)
-
-    safe_plugin_id = _sanitize_optional_plugin_id(request.plugin_id)
-
-    registry = get_plugin_registry()
-    errors = await asyncio.to_thread(
-        registry.install_from_git,
-        request.repository,
-        plugin_id=safe_plugin_id,
-        branch=safe_branch,
-    )
-
-    if errors:
-        raise HTTPException(status_code=400, detail="; ".join(errors))
-
-    # Derive the final plugin id
-    pid = safe_plugin_id
-    if pid is None:
-        from .plugins.sources import plugin_id_from_repo_name, repo_name_from_url
-
-        pid = plugin_id_from_repo_name(repo_name_from_url(request.repository))
-
-    return {
-        "status": "success",
-        "plugin_id": pid,
-        "message": f"Plugin '{pid}' installed from {request.repository}.",
-    }
-
-
-@app.delete("/plugins/{plugin_id}/uninstall")
-async def uninstall_external_plugin(plugin_id: str):
-    """
-    Uninstall an external (non-built-in) plugin.
-
-    Built-in plugins shipped with FiestaBoard cannot be uninstalled.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-
-    # Collect instance compound keys before uninstall so we can purge their configs
-    instance_keys = [
-        p["id"] for p in registry.list_plugins() if p.get("base_plugin_id") == plugin_id and p.get("instance_label")
-    ]
-
-    errors = registry.uninstall_external_plugin(plugin_id)
-
-    if errors:
-        raise HTTPException(status_code=400, detail="; ".join(errors))
-
-    # Purge persisted configs for both the base plugin and every named instance.
-    # The base-id delete is critical: without it the v2→v3 auto-migration would
-    # see the leftover entry as orphaned on the next boot and silently reinstall
-    # the plugin the user just deleted (issue #937).
-    config_manager = get_config_manager()
-    for compound_key in instance_keys:
-        config_manager.delete_plugin_config(compound_key)
-    config_manager.delete_plugin_config(plugin_id)
-
-    return {
-        "status": "success",
-        "plugin_id": plugin_id,
-        "message": f"Plugin '{plugin_id}' has been uninstalled.",
-    }
-
-
-@app.post("/plugins/updates/check")
-async def trigger_plugin_update_check():
-    """
-    Trigger an immediate update check for all external plugins.
-
-    Runs ``git ls-remote`` against each external plugin's origin in a thread
-    pool so the event loop is not blocked.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    loop = asyncio.get_event_loop()
-    results = await loop.run_in_executor(None, registry.check_for_updates)
-    plugins_with_updates = [pid for pid, has_update in results.items() if has_update]
-    return {
-        "checked": len(results),
-        "updates_available": plugins_with_updates,
-    }
-
-
-@app.post("/plugins/{plugin_id}/update")
-async def update_plugin(plugin_id: str):
-    """
-    Fetch the latest commits for an external plugin from its remote and reload it.
-
-    Built-in plugins cannot be updated via this endpoint.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    source = registry.get_plugin_source(plugin_id)
-
-    if source is None:
-        raise HTTPException(status_code=404, detail=f"Plugin '{plugin_id}' not found.")
-
-    if source.source_type == "builtin":
-        raise HTTPException(
-            status_code=400,
-            detail=f"Plugin '{plugin_id}' is a built-in plugin and cannot be updated this way.",
-        )
-
-    if not source.local_path:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Plugin '{plugin_id}' has no local path for updating.",
-        )
-
-    import os as _os
-
-    from .plugins.sources import clone_or_update_repo, get_external_plugins_dir
-
-    # Verify the plugin's local_path is within the external plugins directory
-    # before updating, as a defence-in-depth check.
-    _ext_dir = get_external_plugins_dir()
-    _ext_root = _os.path.realpath(str(_ext_dir))
-    _real_local = _os.path.realpath(str(source.local_path))
-    try:
-        _common = _os.path.commonpath([_ext_root, _real_local])
-    except ValueError:
-        raise HTTPException(status_code=400, detail="Invalid plugin path.") from None
-    if _common != _ext_root or _real_local == _ext_root:
-        raise HTTPException(status_code=400, detail="Invalid plugin path.")
-
-    if not (_real_local and (Path(_real_local) / ".git").is_dir()):
-        raise HTTPException(
-            status_code=400,
-            detail=f"Plugin '{plugin_id}' is not a git repository.",
-        )
-
-    # Pass the validated plugin_id — clone_or_update_repo resolves the path
-    # internally so no user-controlled Path flows into subprocess sinks.
-    # Both the git fetch and the module reimport go to a worker thread so the
-    # event loop keeps serving other requests during the update (#1750).
-    ok, err = await asyncio.to_thread(clone_or_update_repo, "", plugin_id, external_dir=_ext_dir)
-    if not ok:
-        raise HTTPException(status_code=500, detail=f"Update failed: {err}")
-
-    reloaded = await asyncio.to_thread(registry.reload_plugin, plugin_id)
-    if reloaded is None:
-        errors = registry.get_load_errors().get(plugin_id, [])
-        detail = "; ".join(errors) if errors else "Plugin failed to reload after update."
-        raise HTTPException(status_code=500, detail=detail)
-
-    registry._update_status.pop(plugin_id, None)
-
-    return {
-        "status": "success",
-        "plugin_id": plugin_id,
-        "message": f"Plugin '{plugin_id}' has been updated and reloaded.",
-    }
-
-
-@app.post("/plugins/updates/apply")
-async def apply_all_plugin_updates():
-    """
-    Fetch and reload all external plugins that have a pending update.
-
-    Uses the cached update status from the last check — call
-    ``POST /plugins/updates/check`` first if you want a fresh scan before
-    applying.  Returns 200 even when some plugins fail so the caller can
-    inspect partial results.
-    """
-    if not PLUGIN_SYSTEM_AVAILABLE:
-        raise HTTPException(status_code=503, detail="Plugin system is not available.")
-
-    registry = get_plugin_registry()
-    pending = [pid for pid, has_update in registry.get_update_status().items() if has_update]
-
-    if not pending:
-        return {"updated": [], "failed": {}, "message": "No updates available."}
-
-    import os as _os
-    from pathlib import Path as _Path
-
-    from .plugins.sources import clone_or_update_repo, get_external_plugins_dir
-
-    updated: list = []
-    failed: dict = {}
-    _ext_dir = get_external_plugins_dir()
-    _ext_root = _os.path.realpath(str(_ext_dir))
-
-    for plugin_id in pending:
-        source = registry.get_plugin_source(plugin_id)
-        if source is None or not source.local_path:
-            failed[plugin_id] = "Plugin source not found."
-            continue
-
-        _real_local = _os.path.realpath(str(_Path(source.local_path)))
-        try:
-            _common = _os.path.commonpath([_ext_root, _real_local])
-        except ValueError:
-            failed[plugin_id] = "Invalid plugin path."
-            continue
-        if _common != _ext_root or _real_local == _ext_root:
-            failed[plugin_id] = "Invalid plugin path."
-            continue
-        if not (_Path(_real_local) / ".git").is_dir():
-            failed[plugin_id] = "Plugin is not a git repository."
-            continue
-
-        # Pass the validated plugin_id — clone_or_update_repo resolves the path
-        # internally so no user-controlled Path flows into subprocess sinks.
-        # A bulk update is N sequential git fetches; keeping them on the loop
-        # would freeze the API for the sum of all of them (#1750).
-        ok, err = await asyncio.to_thread(clone_or_update_repo, "", plugin_id, external_dir=_ext_dir)
-        if not ok:
-            failed[plugin_id] = f"git fetch failed: {err}"
-            continue
-
-        reloaded = await asyncio.to_thread(registry.reload_plugin, plugin_id)
-        if reloaded is None:
-            errors = registry.get_load_errors().get(plugin_id, [])
-            failed[plugin_id] = "; ".join(errors) if errors else "Reload failed."
-            continue
-
-        registry._update_status.pop(plugin_id, None)
-        updated.append(plugin_id)
-        logger.info("Bulk update: applied update for plugin '%s'", plugin_id)
-
-    return {
-        "updated": updated,
-        "failed": failed,
-        "message": f"Updated {len(updated)} plugin(s); {len(failed)} failed.",
-    }
-
+# =============================================================================
+# Plugin Endpoints — moved to src/plugins/routes.py (issue #1757)
+# =============================================================================
+
+from .plugins.routes import router as plugins_router  # noqa: E402
+
+app.include_router(plugins_router)
+
+# Compatibility aliases: src/mcp_server.py still imports these plugin handler
+# names from api_server. The #1757 MCP commit repoints those tools at
+# PluginService and deletes this block.
+from .plugins.routes import (  # noqa: E402, F401
+    PluginConfigRequest,
+    disable_plugin,
+    enable_plugin,
+    install_registry_plugin,
+    uninstall_external_plugin,
+    update_plugin,
+    update_plugin_config,
+)
 
 # =============================================================================
 # Triggers — Event-based plugin messages
